+++ lib/configuration/__init__.py
@@ -190,6 +190,9 @@ def validate_config_against_methods(config):
 
     exit(1)
 
+  if config['rootFolder'][-1:] == '/':
+    config['rootFolder'] = config['rootFolder'][:-1]
+
 def get_default_config_from_methods(config, settings, defaults):
   from lib import methods
 
@@ -246,7 +249,8 @@ def get_configuration(name):
       'supportsZippedBackups': True,
       'tmpFolder': '/tmp',
       'scripts': {},
-      'executables': config['executables']
+      'executables': config['executables'],
+      'runsLocally': False
     }
 
     defaults = get_default_config_from_methods(host_config, config, defaults)